requirements.txt
requests
requests-toolbelt
plotly
Pillow
